+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - If using cookie-based tracking, the cookie lifetime has been changed to expire at midnight (so a maximum of 24 hours).
 - Simplified client-side tracking script so it's now smaller than 500 bytes.
 - Excluded IP addresses now work properly with the optimized endpoint.
+- Added new dashboard date preset for "all time".
 
 
 ### 1.7.4 - May 14, 2025